test: mirror unit test layout
