+++ tests/unit/api/domains/cho_test.py
@@ -1,21 +1,12 @@
 from __future__ import annotations
 
-import io
 from datetime import date
 
 import pytest
-from starlette.datastructures import FormData
-from starlette.datastructures import UploadFile
 
 from app.api.domains import cho
-from app.api.domains import osu
-from app.commands import ParsingError
-from app.commands import parse__with__command_args
-from app.commands import status_to_id
-from app.constants.mods import Mods
 from app.objects.player import OsuStream
 from app.packets import MultiplayerMatch
-from app.usecases import score_submission as score_submission_usecase
 
 
 def test_parse_login_data_handles_protocol_trailing_newline() -> None:
@@ -107,95 +98,3 @@ def test_validate_match_data_rejects_untrusted_fields(
     match_data.name = name
 
     assert cho.validate_match_data(match_data, expected_host_id) is False
-
-
-def test_parse_score_form_data_returns_score_bytes_and_replay_file() -> None:
-    replay_file = UploadFile(filename="score.osr", file=io.BytesIO(b"replay"))
-    form_data = FormData(
-        [
-            ("score", "encrypted-score"),
-            ("score", replay_file),
-        ],
-    )
-
-    parsed = osu.parse_form_data_score_params(form_data)
-
-    assert parsed == (b"encrypted-score", replay_file)
-
-
-@pytest.mark.parametrize(
-    "form_data",
-    [
-        FormData([("score", "encrypted-score")]),
-        FormData(
-            [
-                ("score", "encrypted-score"),
-                ("score", "not-a-file"),
-            ],
-        ),
-    ],
-)
-def test_parse_score_form_data_rejects_invalid_score_parts(
-    form_data: FormData,
-) -> None:
-    assert osu.parse_form_data_score_params(form_data) is None
-
-
-def test_score_submission_chart_entry_formats_optional_before_and_after_values() -> (
-    None
-):
-    assert score_submission_usecase.chart_entry("rankedScore", None, 123.45) == (
-        "rankedScoreBefore:|rankedScoreAfter:123.45"
-    )
-
-
-def test_score_submission_achievement_string_uses_client_delimiters() -> None:
-    assert (
-        score_submission_usecase.format_achievement_string(
-            "osu-combo-500",
-            "500 Combo",
-            "Achieve a 500 combo.",
-        )
-        == "osu-combo-500+500 Combo+Achieve a 500 combo."
-    )
-
-
-def test_with_command_arg_parser_accepts_acc_misses_combo_and_mods() -> None:
-    parsed = parse__with__command_args(0, ["95.5%", "1m", "429x", "+hddt"])
-
-    assert parsed == {
-        "acc": 95.5,
-        "mods": Mods.HIDDEN | Mods.DOUBLETIME,
-        "combo": 429,
-        "nmiss": 1,
-    }
-
-
-@pytest.mark.parametrize(
-    ("args", "expected_error"),
-    [
-        ([], "Invalid syntax: !with <acc/nmiss/combo/mods ...>"),
-        (["101%"], "Invalid accuracy."),
-        (["bad"], "Unknown argument: bad"),
-    ],
-)
-def test_with_command_arg_parser_rejects_invalid_args(
-    args: list[str],
-    expected_error: str,
-) -> None:
-    parsed = parse__with__command_args(0, args)
-
-    assert isinstance(parsed, ParsingError)
-    assert str(parsed) == expected_error
-
-
-@pytest.mark.parametrize(
-    ("status", "expected_id"),
-    [
-        ("unrank", 0),
-        ("rank", 2),
-        ("love", 5),
-    ],
-)
-def test_status_to_id(status: str, expected_id: int) -> None:
-    assert status_to_id(status) == expected_id
